# packages/api-client/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "ReCredita API",
    "description": "Recuperação (Plano A) e aquisição (Plano B) de veículos.",
    "version": "0.1.0"
  },
  "components": {
    "schemas": {
...
      "Erro": {
        "type": "object",
        "properties": {
          "erro": {
            "type": "string"
          }
        },
        "required": [
          "erro"
        ],
        "additionalProperties": false
      },
...
      "Criado": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          }
        },
        "required": [
          "id"
        ],
        "additionalProperties": false
      },
...
      "Integracao": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "tipo": {
            "type": "string",
            "enum": [
              "apibrasil",
              "evolution"
            ]
          },
          "nome": {
            "type": "string"
          },
          "baseUrl": {
            "type": "string"
          },
          "config": {
            "type": "object",
            "properties": {
              "instancia": {
                "type": "string"
              },
              "servico": {
                "type": "string",
                "enum": [
                  "dados",
                  "consulta"
                ]
              }
            },
            "additionalProperties": false
          },
          "bureauId": {
            "nullable": true,
            "type": "string"
          },
          "bureauNome": {
            "nullable": true,
            "type": "string"
          },
          "custoConsulta": {
            "nullable": true,
            "type": "number"
          },
          "contrato": {
            "nullable": true,
            "type": "string"
          },
          "ativo": {
            "type": "boolean"
          },
          "credenciaisFinal": {
            "type": "string"
          },
          "webhookConfigurado": {
            "type": "boolean"
          },
          "ultimoTesteEm": {
            "nullable": true,
            "type": "string",
            "format": "date-time"
          },
          "ultimoTesteOk": {
            "nullable": true,
            "type": "boolean"
          },
          "ultimoTesteDetalhe": {
            "nullable": true,
            "type": "string"
          },
          "criadoEm": {
            "type": "string",
            "format": "date-time"
          }
        },
        "required": [
          "id",
          "tipo",
          "nome",
          "baseUrl",
          "config",
          "bureauId",
          "bureauNome",
          "custoConsulta",
          "contrato",
          "ativo",
          "credenciaisFinal",
          "webhookConfigurado",
          "ultimoTesteEm",
          "ultimoTesteOk",
          "ultimoTesteDetalhe",
          "criadoEm"
        ],
        "additionalProperties": false
      },
      "IntegracaoCriada": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "webhookUrl": {
            "nullable": true,
            "type": "string"
          }
        },
        "required": [
          "id",
          "webhookUrl"
        ],
        "additionalProperties": false
      },
      "TesteIntegracao": {
        "type": "object",
        "properties": {
          "ok": {
            "type": "boolean"
          },
          "detalhe": {
            "type": "string"
          }
        },
        "required": [
          "ok",
          "detalhe"
        ],
        "additionalProperties": false
      },
      "WebhookConfigurado": {
        "type": "object",
        "properties": {
          "url": {
            "type": "string"
          },
          "configuradoNaEvolution": {
            "type": "boolean"
          },
          "detalhe": {
            "type": "string"
          }
        },
        "required": [
          "url",
          "configuradoNaEvolution",
          "detalhe"
        ],
        "additionalProperties": false
      },
...
    }
  },
  "paths": {
...
    "/api/integracoes": {
      "get": {
        "tags": [
          "integracoes"
        ],
        "responses": {
          "200": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Integracao"
                  }
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "integracoes"
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "oneOf": [
                  {
                    "type": "object",
                    "properties": {
                      "tipo": {
                        "type": "string",
                        "enum": [
                          "apibrasil"
                        ]
                      },
                      "nome": {
                        "type": "string",
                        "minLength": 2
                      },
                      "baseUrl": {
                        "default": "https://gateway.apibrasil.io/api/v2",
                        "type": "string",
                        "format": "uri"
                      },
                      "bearerToken": {
                        "type": "string",
                        "minLength": 10
                      },
                      "deviceToken": {
                        "type": "string",
                        "minLength": 4
                      },
                      "servico": {
                        "default": "dados",
                        "type": "string",
                        "enum": [
                          "dados",
                          "consulta"
                        ]
                      },
                      "contratoFornecedorId": {
                        "type": "string",
                        "minLength": 3
                      },
                      "custoConsulta": {
                        "default": 0,
                        "type": "number",
                        "minimum": 0
                      }
                    },
                    "required": [
                      "tipo",
                      "nome",
                      "bearerToken",
                      "contratoFornecedorId"
                    ],
                    "additionalProperties": false
                  },
                  {
                    "type": "object",
                    "properties": {
                      "tipo": {
                        "type": "string",
                        "enum": [
                          "evolution"
                        ]
                      },
                      "nome": {
                        "type": "string",
                        "minLength": 2
                      },
                      "baseUrl": {
                        "type": "string",
                        "format": "uri"
                      },
                      "apikey": {
                        "type": "string",
                        "minLength": 8
                      },
                      "instancia": {
                        "type": "string",
                        "minLength": 1
                      }
                    },
                    "required": [
                      "tipo",
                      "nome",
                      "baseUrl",
                      "apikey",
                      "instancia"
                    ],
                    "additionalProperties": false
                  }
                ]
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/IntegracaoCriada"
                }
              }
            }
          },
          "409": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Erro"
                }
              }
            }
          }
        }
      }
    },
    "/api/integracoes/{id}": {
      "put": {
        "tags": [
          "integracoes"
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "nome": {
                    "type": "string",
                    "minLength": 2
                  },
                  "baseUrl": {
                    "type": "string",
                    "format": "uri"
                  },
                  "ativo": {
                    "type": "boolean"
                  },
                  "bearerToken": {
                    "type": "string",
                    "minLength": 10
                  },
                  "deviceToken": {
                    "type": "string",
                    "minLength": 4
                  },
                  "apikey": {
                    "type": "string",
                    "minLength": 8
                  },
                  "instancia": {
                    "type": "string",
                    "minLength": 1
                  },
                  "servico": {
                    "type": "string",
                    "enum": [
                      "dados",
                      "consulta"
                    ]
                  }
                },
                "additionalProperties": false
              }
            }
          }
        },
        "parameters": [
          {
            "schema": {
              "type": "string"
            },
            "in": "path",
            "name": "id",
            "required": true
          }
        ],
        "responses": {
          "200": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Criado"
                }
              }
            }
          },
          "404": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Erro"
                }
              }
            }
          }
        }
      }
    },
    "/api/integracoes/{id}/testar": {
      "post": {
        "tags": [
          "integracoes"
        ],
        "parameters": [
          {
            "schema": {
              "type": "string"
            },
            "in": "path",
            "name": "id",
            "required": true
          }
        ],
        "responses": {
          "200": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TesteIntegracao"
                }
              }
            }
          },
          "404": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Erro"
                }
              }
            }
          }
        }
      }
    },
    "/api/integracoes/{id}/webhook": {
      "post": {
        "tags": [
          "integracoes"
        ],
        "parameters": [
          {
            "schema": {
              "type": "string"
            },
            "in": "path",
            "name": "id",
            "required": true
          }
        ],
        "responses": {
          "200": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/WebhookConfigurado"
                }
              }
            }
          },
          "404": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Erro"
                }
              }
            }
          },
          "502": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Erro"
                }
              }
            }
          }
        }
      }
    },
    "/api/casos/{id}/consulta-apibrasil": {
      "post": {
        "tags": [
          "integracoes"
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "integracaoId": {
                    "type": "string",
                    "minLength": 1
                  },
                  "baseLegal": {
                    "type": "string",
                    "enum": [
                      "execucao_contrato",
                      "legitimo_interesse",
                      "obrigacao_legal",
                      "consentimento"
                    ]
                  },
                  "justificativa": {
                    "type": "string",
                    "minLength": 10
                  },
                  "papelPessoa": {
                    "type": "string",
                    "enum": [
                      "devedor",
                      "proprietario",
                      "terceiro_possuidor",
                      "avalista",
                      "outro"
                    ]
                  }
                },
                "required": [
                  "baseLegal",
                  "justificativa"
                ],
                "additionalProperties": false
              }
            }
          }
        },
        "parameters": [
          {
            "schema": {
              "type": "string"
            },
            "in": "path",
            "name": "id",
            "required": true
          }
        ],
        "responses": {
          "201": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ResumoImportacao"
                }
              }
            }
          },
          "404": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Erro"
                }
              }
            }
          },
          "409": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Erro"
                }
              }
            }
          },
          "422": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Erro"
                }
              }
            }
          },
          "502": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Erro"
                }
              }
            }
          }
        }
      }
    },
...
  }
}